Refactor: Renaming / pathing
